+++ mech/mech.py
@@ -242,20 +242,20 @@ def provision(ctx, instance, show_only):
 
     Notes:
 
-    There are a few provision types: 'file', 'shell', and 'pyinfra'.
+    Provision types are: 'file', 'shell' (Bash), 'ps' (Powershell), and 'pyinfra'.
 
-    'shell' can be inline.
+    'shell' or 'ps' can be inline.
 
-    'shell' and 'pyinfra' can have a remote endpoint ('http', 'https', 'ftp') for the script.
+    'shell', 'ps', and 'pyinfra' can have a remote endpoint ('http', 'https', 'ftp') for the script.
     (ex: 'http://example.com/somefile.sh' or ex: 'ftp://foo.com/install.sh')
 
     'pyinfra' scripts must end with '.py' and 'pyinfra' must be installed.
     See https://pyinfra.readthedocs.io/en/latest/ for more info.
 
-    Provisioning is run when the instance is started.
-    This option is if you want to re-run the provisioning.
+    Provisioning is run when the instance is started or you can re-run the provisioning.
 
-    An example of provisioning could be installing puppet (or your config tool of choice).
+    An example of provisioning could be installing puppet (or your config tool of choice)
+    or preparing the instance "just the way you want it".
 
     '''